PNG Guides and EmbPattern init.
PngWriter has guides added.
A couple corrections to PngWriter.
PyEmbroidery is turned into a shell references to EmbPattern static calls.
EmbPattern has init parameters of filename, or EmbPattern.
EmbPattern has .read() and .write() which accept settings as direct keyword arguments.
