+++ pyembroidery/PngWriter.py
@@ -7,11 +7,181 @@ from .EmbThread import EmbThread
 SEQUIN_CONTINGENCY = CONTINGENCY_SEQUIN_STITCH
 FULL_JUMP = True
 
+characters = {
+    '0': [[9, 9, 4, 1, 0, 1, 5, 9, 9],
+          [9, 3, 0, 0, 0, 0, 0, 5, 9],
+          [7, 0, 1, 7, 9, 6, 0, 0, 9],
+          [4, 0, 5, 9, 9, 9, 4, 0, 6],
+          [3, 0, 8, 9, 9, 9, 6, 0, 4],
+          [2, 0, 9, 5, 0, 6, 8, 0, 3],
+          [2, 0, 9, 4, 0, 6, 8, 0, 3],
+          [2, 0, 9, 9, 9, 9, 8, 0, 3],
+          [3, 0, 8, 9, 9, 9, 6, 0, 4],
+          [4, 0, 5, 9, 9, 9, 4, 0, 6],
+          [7, 0, 1, 7, 9, 6, 0, 0, 8],
+          [9, 3, 0, 0, 0, 0, 0, 5, 9],
+          [9, 9, 4, 1, 0, 1, 5, 9, 9]],
+
+    '1': [[9, 8, 5, 2, 0, 1, 9, 9, 9],
+          [9, 1, 0, 0, 0, 1, 9, 9, 9],
+          [9, 2, 4, 7, 1, 1, 9, 9, 9],
+          [9, 9, 9, 9, 1, 1, 9, 9, 9],
+          [9, 9, 9, 9, 1, 1, 9, 9, 9],
+          [9, 9, 9, 9, 1, 1, 9, 9, 9],
+          [9, 9, 9, 9, 1, 1, 9, 9, 9],
+          [9, 9, 9, 9, 1, 1, 9, 9, 9],
+          [9, 9, 9, 9, 1, 1, 9, 9, 9],
+          [9, 9, 9, 9, 1, 1, 9, 9, 9],
+          [9, 9, 9, 9, 1, 1, 9, 9, 9],
+          [9, 3, 0, 0, 0, 0, 0, 0, 3],
+          [9, 3, 0, 0, 0, 0, 0, 0, 3]],
+
+    '2': [[8, 5, 2, 1, 1, 2, 6, 9, 9],
+          [4, 0, 0, 0, 0, 0, 0, 4, 9],
+          [5, 4, 7, 8, 8, 5, 0, 0, 8],
+          [9, 9, 9, 9, 9, 9, 3, 0, 7],
+          [9, 9, 9, 9, 9, 9, 3, 0, 9],
+          [9, 9, 9, 9, 9, 8, 0, 4, 9],
+          [9, 9, 9, 9, 8, 1, 2, 9, 9],
+          [9, 9, 9, 8, 2, 2, 9, 9, 9],
+          [9, 9, 8, 1, 2, 9, 9, 9, 9],
+          [9, 8, 1, 2, 9, 9, 9, 9, 9],
+          [7, 1, 2, 9, 9, 9, 9, 9, 9],
+          [3, 0, 0, 0, 0, 0, 0, 0, 6],
+          [3, 0, 0, 0, 0, 0, 0, 0, 6]],
+
+    '3': [[9, 6, 2, 1, 1, 2, 6, 9, 9],
+          [5, 0, 0, 0, 0, 0, 0, 4, 9],
+          [6, 3, 7, 8, 9, 6, 1, 0, 8],
+          [9, 9, 9, 9, 9, 9, 3, 0, 8],
+          [9, 9, 9, 9, 8, 6, 1, 1, 9],
+          [9, 9, 4, 0, 0, 0, 2, 8, 9],
+          [9, 9, 4, 0, 0, 0, 2, 8, 9],
+          [9, 9, 9, 9, 8, 6, 1, 1, 8],
+          [9, 9, 9, 9, 9, 9, 5, 0, 6],
+          [9, 9, 9, 9, 9, 9, 5, 0, 5],
+          [3, 5, 7, 9, 8, 6, 1, 0, 7],
+          [2, 0, 0, 0, 0, 0, 0, 3, 9],
+          [8, 4, 2, 0, 1, 2, 5, 9, 9]],
+
+    '4': [[9, 9, 9, 9, 7, 0, 0, 7, 9],
+          [9, 9, 9, 9, 2, 1, 0, 7, 9],
+          [9, 9, 9, 5, 2, 4, 0, 7, 9],
+          [9, 9, 8, 1, 7, 4, 0, 7, 9],
+          [9, 9, 3, 4, 9, 4, 0, 7, 9],
+          [9, 6, 1, 9, 9, 4, 0, 7, 9],
+          [8, 1, 6, 9, 9, 4, 0, 7, 9],
+          [4, 3, 9, 9, 9, 4, 0, 7, 9],
+          [0, 0, 0, 0, 0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0, 0, 0, 0, 0],
+          [9, 9, 9, 9, 9, 4, 0, 7, 9],
+          [9, 9, 9, 9, 9, 4, 0, 7, 9],
+          [9, 9, 9, 9, 9, 4, 0, 7, 9]],
+
+    '5': [[7, 0, 0, 0, 0, 0, 0, 5, 9],
+          [7, 0, 0, 0, 0, 0, 0, 5, 9],
+          [7, 0, 5, 9, 9, 9, 9, 9, 9],
+          [7, 0, 5, 9, 9, 9, 9, 9, 9],
+          [7, 0, 1, 1, 0, 2, 6, 9, 9],
+          [7, 0, 0, 0, 0, 0, 0, 5, 9],
+          [7, 4, 7, 9, 8, 4, 0, 0, 9],
+          [9, 9, 9, 9, 9, 9, 3, 0, 6],
+          [9, 9, 9, 9, 9, 9, 5, 0, 6],
+          [9, 9, 9, 9, 9, 9, 3, 0, 6],
+          [3, 5, 8, 9, 8, 5, 0, 0, 9],
+          [2, 0, 0, 0, 0, 0, 0, 6, 9],
+          [8, 4, 1, 0, 1, 2, 7, 9, 9]],
+
+    '6': [[9, 9, 7, 2, 1, 0, 3, 8, 9],
+          [9, 6, 0, 0, 0, 0, 0, 3, 9],
+          [8, 0, 0, 4, 8, 9, 6, 4, 9],
+          [5, 0, 4, 9, 9, 9, 9, 9, 9],
+          [3, 0, 8, 9, 9, 9, 9, 9, 9],
+          [2, 1, 8, 3, 0, 1, 3, 8, 9],
+          [2, 2, 2, 0, 0, 0, 0, 1, 8],
+          [2, 0, 2, 7, 9, 8, 2, 0, 5],
+          [3, 0, 7, 9, 9, 9, 7, 0, 3],
+          [4, 0, 7, 9, 9, 9, 7, 0, 3],
+          [7, 0, 2, 7, 9, 8, 2, 0, 5],
+          [9, 3, 0, 0, 0, 0, 0, 1, 8],
+          [9, 9, 4, 1, 0, 1, 3, 8, 9]],
+
+    '7': [[2, 0, 0, 0, 0, 0, 0, 0, 5],
+          [2, 0, 0, 0, 0, 0, 0, 0, 7],
+          [9, 9, 9, 9, 9, 9, 3, 2, 9],
+          [9, 9, 9, 9, 9, 8, 0, 5, 9],
+          [9, 9, 9, 9, 9, 4, 0, 8, 9],
+          [9, 9, 9, 9, 9, 1, 3, 9, 9],
+          [9, 9, 9, 9, 6, 0, 6, 9, 9],
+          [9, 9, 9, 9, 2, 1, 9, 9, 9],
+          [9, 9, 9, 7, 0, 4, 9, 9, 9],
+          [9, 9, 9, 3, 0, 8, 9, 9, 9],
+          [9, 9, 8, 0, 2, 9, 9, 9, 9],
+          [9, 9, 5, 0, 6, 9, 9, 9, 9],
+          [9, 9, 1, 0, 9, 9, 9, 9, 9]],
+
+    '8': [[9, 8, 3, 1, 0, 1, 4, 8, 9],
+          [8, 1, 0, 0, 0, 0, 0, 1, 9],
+          [5, 0, 3, 8, 9, 7, 2, 0, 6],
+          [5, 0, 7, 9, 9, 9, 5, 0, 6],
+          [7, 0, 3, 8, 9, 7, 2, 1, 8],
+          [9, 7, 2, 0, 0, 0, 2, 7, 9],
+          [9, 5, 1, 0, 0, 0, 1, 6, 9],
+          [6, 0, 3, 8, 9, 7, 2, 0, 7],
+          [2, 0, 9, 9, 9, 9, 7, 0, 4],
+          [2, 0, 9, 9, 9, 9, 7, 0, 3],
+          [3, 0, 3, 8, 9, 7, 2, 0, 4],
+          [7, 0, 0, 0, 0, 0, 0, 1, 8],
+          [9, 8, 3, 1, 0, 1, 4, 8, 9]],
+
+    '9': [[9, 7, 3, 0, 0, 2, 6, 9, 9],
+          [7, 0, 0, 0, 0, 0, 0, 5, 9],
+          [3, 0, 4, 8, 9, 7, 1, 0, 9],
+          [1, 0, 9, 9, 9, 9, 5, 0, 6],
+          [1, 1, 9, 9, 9, 9, 5, 0, 5],
+          [3, 0, 4, 8, 9, 7, 1, 0, 4],
+          [7, 0, 0, 0, 0, 0, 3, 0, 4],
+          [9, 7, 2, 0, 1, 4, 7, 0, 4],
+          [9, 9, 9, 9, 9, 9, 5, 0, 5],
+          [9, 9, 9, 9, 9, 9, 2, 0, 7],
+          [9, 3, 7, 9, 8, 3, 0, 2, 9],
+          [9, 1, 0, 0, 0, 0, 0, 7, 9],
+          [9, 7, 2, 0, 1, 3, 8, 9, 9]],
+
+    '-': [[9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 1, 0, 0, 0, 3, 9, 9],
+          [9, 9, 1, 0, 0, 0, 3, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9]],
+
+    'm': [[9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [9, 9, 9, 9, 9, 9, 9, 9, 9],
+          [0, 5, 2, 1, 7, 6, 1, 2, 8],
+          [0, 1, 0, 0, 1, 1, 0, 0, 3],
+          [0, 2, 9, 4, 0, 4, 8, 2, 1],
+          [0, 5, 9, 6, 0, 7, 9, 4, 1],
+          [0, 5, 9, 6, 0, 7, 9, 4, 0],
+          [0, 5, 9, 6, 0, 7, 9, 4, 0],
+          [0, 5, 9, 6, 0, 7, 9, 4, 0],
+          [0, 5, 9, 6, 0, 7, 9, 4, 0],
+          [0, 5, 9, 6, 0, 7, 9, 4, 0],
+          [0, 5, 9, 6, 0, 7, 9, 4, 0]]
+}
+
 
 def write_png(buf, width, height):
     width_byte_4 = width * 4
     raw_data = b"".join(
-        b'\x00' + buf[span:span + width_byte_4] for span in range((height - 1) * width * 4, -1, - width_byte_4))
+        b'\x00' + buf[span:span + width_byte_4] for span in range(0, height * width * 4, width_byte_4))
 
     def png_pack(png_tag, data):
         chunk_head = png_tag + data
@@ -105,8 +275,43 @@ class PngBuffer:
             for pos in range(-left, right):
                 self.plot(x, y + pos)
 
+    def draw_text(self, x, y, string, rotate=False):
+        for c in string:
+            m = characters[c]
+            for cx in range(len(m[0])):
+                for cy in range(len(m)):
+                    v = m[cy][cx]
+                    if v == 9:
+                        continue
+                    if rotate:
+                        gx = x + (len(m) - cy) + 1
+                        gy = y + cx + 2
+                    else:
+                        gx = x + cx + 2
+                        gy = y + cy + 2
+                    pos = (self.width * gy) + gx
+                    idx = pos * 4
+                    a2 = (9.0 - v) / 9.0
+                    r = (1.0 - a2) * self.buf[idx]
+                    g = (1.0 - a2) * self.buf[idx + 1]
+                    b = (1.0 - a2) * self.buf[idx + 2]
+                    a1 = self.buf[idx + 3] / 255.0
+                    a = a2 + a1 * (1.0 - a2)
+                    try:
+                        self.buf[idx] = int(r)
+                        self.buf[idx + 1] = int(g)
+                        self.buf[idx + 2] = int(b)
+                        self.buf[idx + 3] = int(a * 255)
+                    except IndexError:
+                        pass
+            if rotate:
+                y += 11
+            else:
+                x += 11
+
 
 def write(pattern, f, settings=None):
+    guides = settings.get("guides", False)
     extends = pattern.bounds()
     pattern.translate(-extends[0], -extends[1])
     width = int(extends[2] - extends[0])
@@ -133,9 +338,31 @@ def write(pattern, f, settings=None):
         last_y = None
         for stitch in block:
             x = int(stitch[0])
-            y = -int(stitch[1])
+            y = int(stitch[1])
             if last_x is not None:
                 draw_buff.draw_line(last_x, last_y, x, y)
             last_x = x
             last_y = y
+
+    if guides:
+        draw_buff.red = 0
+        draw_buff.green = 0
+        draw_buff.blue = 0
+        draw_buff.alpha = 255
+        draw_buff.line_width = 1
+        min_x = extends[0]
+        min_y = extends[1]
+        points = 50
+        draw_buff.draw_text(0, 0, 'mm')
+        for x in range(points-(min_x % points), width - 30, points):
+            if x < 30:
+                continue
+            draw_buff.draw_text(x, 0, str(int((x + min_x) / 10)), rotate=True)
+            draw_buff.draw_line(x, 0, x, 30)
+        for y in range(points-(min_y % points), height - 30, points):
+            if y < 30:
+                continue
+            draw_buff.draw_text(0, y, str(int((y + min_y) / 10)))
+            draw_buff.draw_line(0, y, 30, y)
+
     f.write(write_png(bytes(bytearray(draw_buff.buf)), draw_buff.width, draw_buff.height))
